move search query tokenizer to @tutanota/utils
