+++ packages/tutanota-utils/lib/Tokenizer.ts
