+++ packages/tutanota-utils/lib/index.ts
@@ -169,3 +169,5 @@ export {
 export { mod, clamp } from "./MathUtils.js"
 
 export { renderCsv } from "./Csv.js"
+
+export { tokenize } from "./Tokenizer.js"
