+++ packages/tutanota-utils/test/TokenizerTest.ts
@@ -1,5 +1,5 @@
 import o from "@tutao/otest"
-import { tokenize } from "../../../../../src/api/worker/search/Tokenizer.js"
+import { tokenize } from "../lib/Tokenizer.js"
 o.spec("Tokenizer test", () => {
 	o("tokenize", () => {
 		o(tokenize("")).deepEquals([])
