+++ src/api/worker/search/ContactIndexer.ts
@@ -8,7 +8,7 @@ import { neverNull, noOp, ofClass, promiseMap } from "@tutao/tutanota-utils"
 import { FULL_INDEXED_TIMESTAMP, OperationType } from "../../common/TutanotaConstants"
 import { IndexerCore } from "./IndexerCore"
 import { SuggestionFacade } from "./SuggestionFacade"
-import { tokenize } from "./Tokenizer"
+import { tokenize } from "@tutao/tutanota-utils"
 import type { EntityUpdate } from "../../entities/sys/TypeRefs.js"
 import { EntityClient } from "../../common/EntityClient"
 import { GroupDataOS, MetaDataOS } from "./IndexTables.js"
