+++ src/api/worker/search/GroupInfoIndexer.ts
@@ -8,7 +8,7 @@ import type { Db, GroupData, IndexUpdate, SearchIndexEntry } from "./SearchTypes
 import { _createNewIndexUpdate, typeRefToTypeInfo, userIsLocalOrGlobalAdmin } from "./IndexUtils"
 import { IndexerCore } from "./IndexerCore"
 import { SuggestionFacade } from "./SuggestionFacade"
-import { tokenize } from "./Tokenizer"
+import { tokenize } from "@tutao/tutanota-utils"
 import { EntityClient } from "../../common/EntityClient"
 import { GroupDataOS } from "./IndexTables.js"
