+++ test/tests/Suite.ts
@@ -13,7 +13,6 @@ import "./api/common/error/TutanotaErrorTest.js"
 import "./api/worker/rest/EntityRestCacheTest.js"
 import "./api/worker/rest/EphemeralCacheStorageTest.js"
 import "./api/worker/EventBusClientTest.js"
-import "./api/worker/search/TokenizerTest.js"
 import "./api/worker/search/IndexerTest.js"
 import "./api/worker/search/IndexerCoreTest.js"
 import "./api/worker/search/ContactIndexerTest.js"
